requirements.txt
arrow==0.15.4
cdsapi==0.2.3
geopy==1.20.0
matplotlib==3.1.1
MetPy==0.11.1
networkx==2.4
numpy==1.17.3
pandas==0.25.3
Pillow==6.2.1
pyramid-arima==0.9.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.1.2
scikit-image==0.16.2
scikit-learn==0.21.3
scipy==1.3.1
torch==1.1.0
torch-cluster==1.4.5
torch-geometric==1.3.2
torch-scatter==1.4.0
torch-sparse==0.4.3
torch-spline-conv==1.1.1
tqdm==4.38.0
xarray==0.14.0
